Refactored BaseDQN, Noisy and Dueling functional now
