+++ genrl/deep/agents/dqn/__init__.py
@@ -1,5 +1,6 @@
-from genrl.deep.agents.dqn.base import DQN  # noqa
+from genrl.deep.agents.dqn.base import BaseDQN  # noqa
 from genrl.deep.agents.dqn.double import DoubleDQN  # noqa
+from genrl.deep.agents.dqn.dqn import DQN  # noqa
 from genrl.deep.agents.dqn.dueling import DuelingDQN  # noqa
 from genrl.deep.agents.dqn.noisy import NoisyDQN  # noqa
 from genrl.deep.agents.dqn.prioritized import PrioritizedReplayDQN  # noqa
