+++ genrl/deep/agents/dqn/noisy.py
@@ -7,17 +7,18 @@ from torch import optim as opt
 
 from ....environments import VecEnv
 from ...common import PrioritizedBuffer, PushReplayBuffer, get_env_properties, get_model
-from .base import DQN
+from .base import BaseDQN
 
 
-class NoisyDQN(DQN):
+class NoisyDQN(BaseDQN):
     def __init__(
         self,
         network_type: str,
         env: Union[gym.Env, VecEnv],
         batch_size: int = 32,
         gamma: float = 0.99,
-        layers: Tuple = (32, 32),
+        layers: Tuple = (128,),
+        noisy_layers: Tuple = (128, 128),
         lr: float = 0.001,
         replay_size: int = 100,
         max_epsilon: float = 1.0,
@@ -38,6 +39,9 @@ class NoisyDQN(DQN):
             epsilon_decay,
             **kwargs,
         )
+        print("running")
+        self.noisy_layers = noisy_layers
+
         self.empty_logs()
         self.create_model()
 
@@ -49,7 +53,7 @@ class NoisyDQN(DQN):
         input_dim, action_dim, _, _ = get_env_properties(self.env, self.network_type)
 
         self.model = get_model("dv", self.network_type + "noisy")(
-            input_dim, action_dim, self.layers
+            input_dim, action_dim, fc_layers=self.layers, noisy_layers=self.noisy_layers
         )
         self.target_model = deepcopy(self.model)
 
@@ -58,7 +62,7 @@ class NoisyDQN(DQN):
 
     def update_params(self, update_interval: int) -> None:
         for timestep in range(update_interval):
-            loss = self.get_td_loss()
+            loss = self.get_q_loss()
             self.optimizer.zero_grad()
             loss.backward()
             self.optimizer.step()
