+++ genrl/deep/agents/dqn/prioritized.py
@@ -7,10 +7,10 @@ from torch import optim as opt
 
 from ....environments import VecEnv
 from ...common import PrioritizedBuffer, get_env_properties, get_model
-from .base import DQN
+from .base import BaseDQN
 
 
-class PrioritizedReplayDQN(DQN):
+class PrioritizedReplayDQN(BaseDQN):
     def __init__(
         self,
         network_type: str,
